Function implementation moved.
Function implementation moved into .c files. Created dockerfile and jenkinsfile. zig-cache has been removed.
