--- docker-compose.yml
@@ -1,1 +1,0 @@
-version: '3.8'
