--- server/.zig-cache/o/a32fcfc0e475f08356ed8aef4020640b/dependencies.zig
@@ -1,2 +1,0 @@
-pub const packages = struct {};
-pub const root_deps: []const struct { []const u8, []const u8 } = &.{};
